chore: rearrange frontend source files
 - elevate the `src/views` folder up to `views` and rename it to `components`
 - create a separate folder for each component (to keep .css file next to the .ts file in the same folder)
 - gather should-not-be-needed and should-be-in-the-framework code inside a `utils` folder

This brings the project structure closer to what the starter wizard generates.
