+++ frontend/components/dashboard-view/dashboard-view.css
