+++ frontend/connect-client.ts
@@ -1,4 +1,4 @@
 import {ConnectClient} from '@vaadin/flow-frontend/Connect';
-import {reloginOnExpiredSession} from "./relogin-on-expired-session";
+import {reloginOnExpiredSession} from "./utils/relogin-on-expired-session";
 const client = new ConnectClient({prefix: 'connect', middlewares: [reloginOnExpiredSession]});
 export default client;
