--- frontend/styles/README
@@ -1,3 +1,0 @@
-Place your optional style modules in this folder.
-
-See https://vaadin.com/docs/flow/theme/theming-overview.html for getting started on theming Vaadin Flow applications.
