+++ frontend/utils/contact-model.ts
@@ -1,5 +1,5 @@
-import Contact from '../../generated/com/vaadin/tutorial/crm/backend/service/endpoint/Contact';
-import Status from '../../generated/com/vaadin/tutorial/crm/backend/entity/Contact/Status';
+import Contact from '../generated/com/vaadin/tutorial/crm/backend/service/endpoint/Contact';
+import Status from '../generated/com/vaadin/tutorial/crm/backend/entity/Contact/Status';
 
 export class ContactModel implements Contact {
   constructor(
